began work on vec cvt
added the MXCSR register, plus flag wrappers for all of its contents.

changed PerformRoundTrip() to take the rounding mode (rather than using FPU_RC).
this is to allow other code sources to use it independently (e.g. new vec cvt instructions).
also changed std::nearbyintl() to std::nearbyint() (will look better if double/float overloads are added for simd rounding).

added a new typedef: VPUCVTDelegate - a member function pointer that converts a value into a different form.

added a packed vec cvt format (not yet completed).

added several forms of scalar vec cvt (separate from the packed form because it can take cpu register operands, which would be more annoying to code).
the scalar forms typically take an xmm reg, cpu reg, or mem val.
the instructions operating on xmm registers are limited by intel to the usual xmm0-15, which allows them to use a compact 4-bit vec reg specifier.
this means each scalar format only takes a single byte (aside from memory operands and the opcode).

added a bunch of helper functions for performing the conversions.
currently, they perform their functions assuming it works.
desired behavior is to return a special value on some failures, which can be added later.

added the VEC_CVT opcode - uses an 8-bit extended code to select the conversion mechanism.

changed ZMMRegister::uint() to take u64 values instead of int (was previously int because C# likes int a lot more than ulong, especially for array indexing).
changed ZMMRegister_sizecode_wrapper's internal index/sizecode info to u16 instead of i16 (neither can be negative anyway and virtually everything in csx64 is unsigned).

+++ ExeTypes.h
@@ -206,13 +206,13 @@ namespace CSX64
 
 		// -- sizecode access utilities -- //
 
-		inline ZMMRegister_sizecode_wrapper uint(int sizecode, int index);
+		inline ZMMRegister_sizecode_wrapper uint(u64 sizecode, u64 index);
 	};
 	struct ZMMRegister_sizecode_wrapper
 	{
 		ZMMRegister &reg;
-		i16 index;
-		i16 sizecode;
+		u16 index;
+		u16 sizecode;
 
 		inline constexpr operator u64() const
 		{
@@ -240,7 +240,7 @@ namespace CSX64
 		}
 		inline constexpr ZMMRegister_sizecode_wrapper operator=(ZMMRegister_sizecode_wrapper other) { *this = (u64)other; return *this; }
 	};
-	ZMMRegister_sizecode_wrapper ZMMRegister::uint(int sizecode, int index) { return {*this, (i16)index, (i16)sizecode}; }
+	ZMMRegister_sizecode_wrapper ZMMRegister::uint(u64 sizecode, u64 index) { return {*this, (u16)index, (u16)sizecode}; }
 
 	/// <summary>
 	/// Represents a file descriptor used by the <see cref="CSX64"/> processor
